+++ packages/atomic-angular/projects/atomic-angular/src/lib/stencil-generated/atomic-angular.module.ts
@@ -7,7 +7,7 @@ import {APP_INITIALIZER, ModuleWithProviders, NgModule, Provider} from '@angular
 import {
 AtomicAriaLive,
 AtomicAutomaticFacet,
-AtomicAutomaticFacetBuilder,
+AtomicAutomaticFacetGenerator,
 AtomicBreadbox,
 AtomicCategoryFacet,
 AtomicColorFacet,
@@ -102,7 +102,7 @@ defineCustomElements(window);
 const DECLARATIONS = [
 AtomicAriaLive,
 AtomicAutomaticFacet,
-AtomicAutomaticFacetBuilder,
+AtomicAutomaticFacetGenerator,
 AtomicBreadbox,
 AtomicCategoryFacet,
 AtomicColorFacet,