+++ packages/atomic/tailwind.config.js
@@ -97,6 +97,5 @@ module.exports = {
         variants('outlineColor')
       );
     }),
-    require('@tailwindcss/line-clamp'),
   ],
 };