+++ packages/headless/src/controllers/index.ts
@@ -245,6 +245,9 @@ export type {
   BreadcrumbManagerState,
   BreadcrumbManager,
   DeselectableValue,
+  AutomaticFacetBreadcrumb,
+  CoreBreadcrumbManager,
+  CoreBreadcrumbManagerState,
 } from './breadcrumb-manager/headless-breadcrumb-manager';
 export {buildBreadcrumbManager} from './breadcrumb-manager/headless-breadcrumb-manager';
 
@@ -412,11 +415,12 @@ export type {
 export {buildAutomaticFacet} from './facets/automatic-facet/headless-automatic-facet';
 
 export type {
-  AutomaticFacetBuilder,
-  AutomaticFacetBuilderProps,
-  AutomaticFacetBuilderState,
-} from './facets/automatic-facet-builder/headless-automatic-facet-builder';
-export {buildAutomaticFacetBuilder} from './facets/automatic-facet-builder/headless-automatic-facet-builder';
+  AutomaticFacetGenerator,
+  AutomaticFacetGeneratorProps,
+  AutomaticFacetGeneratorState,
+  AutomaticFacetGeneratorOptions,
+} from './facets/automatic-facet-generator/headless-automatic-facet-generator';
+export {buildAutomaticFacetGenerator} from './facets/automatic-facet-generator/headless-automatic-facet-generator';
 
 export type {
   GeneratedAnswer,