+++ packages/quantic/cypress/e2e/default-2/generatedAnswer/generated-answer-selectors.ts
@@ -1,0 +1,56 @@
+import {ComponentSelector, CypressSelector} from '../../common-selectors';
+
+export const generatedAnswerComponent = 'c-quantic-generated-answer';
+
+export interface GeneratedAnswerSelector extends ComponentSelector {
+  generatedAnswerCard: () => CypressSelector;
+  generatedAnswerContent: () => CypressSelector;
+  likeButton: () => CypressSelector;
+  dislikeButton: () => CypressSelector;
+  citations: () => CypressSelector;
+  citationTitle: (index: number) => CypressSelector;
+  citationIndex: (index: number) => CypressSelector;
+  citationLink: (index: number) => CypressSelector;
+  retryButton: () => CypressSelector;
+}
+
+export const GeneratedAnswerSelectors: GeneratedAnswerSelector = {
+  get: () => cy.get(generatedAnswerComponent),
+
+  generatedAnswerCard: () =>
+    GeneratedAnswerSelectors.get().find('[data-cy="generated-answer__card"]'),
+  generatedAnswerContent: () =>
+    GeneratedAnswerSelectors.get().find(
+      '[data-cy="generated-answer__content"]'
+    ),
+  likeButton: () =>
+    GeneratedAnswerSelectors.get().find(
+      '[data-cy="generated-answer__feedback"] [data-cy="feedback__like-button"]'
+    ),
+  dislikeButton: () =>
+    GeneratedAnswerSelectors.get().find(
+      '[data-cy="generated-answer__feedback"] [data-cy="feedback__dislike-button"]'
+    ),
+  citations: () =>
+    GeneratedAnswerSelectors.get().find(
+      '[data-cy="generated-answer__citations"]'
+    ),
+  citationTitle: (index: number) =>
+    GeneratedAnswerSelectors.get()
+      .find('[data-cy="generated-answer__citations"] .citation__title')
+      .eq(index),
+
+  citationIndex: (index: number) =>
+    GeneratedAnswerSelectors.get()
+      .find('[data-cy="generated-answer__citations"] .citation__index')
+      .eq(index),
+
+  citationLink: (index: number) =>
+    GeneratedAnswerSelectors.get()
+      .find('[data-cy="generated-answer__citations"] .citation__badge')
+      .eq(index),
+  retryButton: () =>
+    GeneratedAnswerSelectors.get().find(
+      '[data-cy="generated-answer__retry-button"]'
+    ),
+};