+++ packages/quantic/force-app/main/default/lwc/quanticResultChildren/quanticResultChildren.js
@@ -1,12 +1,11 @@
 import hideRelatedItems from '@salesforce/label/c.quantic_HideRelatedItems';
 import loadRelatedItems from '@salesforce/label/c.quantic_LoadRelatedItems';
 import noRelatedItems from '@salesforce/label/c.quantic_NoRelatedItems';
-import { LightningElement, api } from 'lwc';
+import {LightningElement, api} from 'lwc';
 // @ts-ignore
 import loadingTemplate from './loading.html';
 // @ts-ignore
 import resultChildrenTemplate from './quanticResultChildren.html';
-
 
 /** @typedef {import("coveo").FoldedCollection} FoldedCollection */
 /** @typedef {import("coveo").FoldedResult} FoldedResult */
@@ -65,7 +64,7 @@ export default class QuanticResultChildren extends LightningElement {
   /** @type {boolean} */
   areAllChildResultsLoaded = false;
   /** @type {boolean} */
-  areChildResultsExpanded = false;
+  _areChildResultsExpanded = false;
   /** @type {Array<FoldedResult>} */
   firstChildrenPartition;
 
@@ -84,6 +83,10 @@ export default class QuanticResultChildren extends LightningElement {
     return this?.collection?.moreResultsAvailable;
   }
 
+  get areChildResultsExpanded() {
+    return this._areChildResultsExpanded && !this.areMoreResultsAvailable;
+  }
+
   handleToggleChildResults() {
     if (this.areChildResultsExpanded) {
       this.showLessFoldedResults();
@@ -99,13 +102,13 @@ export default class QuanticResultChildren extends LightningElement {
   }
 
   showLessFoldedResults() {
-    this.areChildResultsExpanded = false;
+    this._areChildResultsExpanded = false;
     this.foldedResultListController.logShowLessFoldedResults();
   }
 
   showMoreFoldedResults() {
-    this.areChildResultsExpanded = true;
-    if (!this.areAllChildResultsLoaded) {
+    this._areChildResultsExpanded = true;
+    if (this.areMoreResultsAvailable) {
       this.loadAllFoldedResults();
     } else {
       this.foldedResultListController.logShowMoreFoldedResults(
@@ -177,7 +180,11 @@ export default class QuanticResultChildren extends LightningElement {
     if (!this.isFirstLevelChildCollection) {
       return false;
     }
-    return this.areAllChildResultsLoaded && !this.moreResultsFound;
+    return (
+      this.areAllChildResultsLoaded &&
+      !this.moreResultsFound &&
+      !this.areMoreResultsAvailable
+    );
   }
 
   render() {