+++ packages/samples/headless-ssr/src/app/generic/components/result-list.tsx
@@ -1,0 +1,30 @@
+import {
+  ResultList as ResultListController,
+  ResultListState,
+} from '@coveo/headless';
+import {useEffect, useState, FunctionComponent} from 'react';
+
+interface ResultListProps {
+  initialState: ResultListState;
+  controller?: ResultListController;
+}
+
+export const ResultList: FunctionComponent<ResultListProps> = (props) => {
+  const {initialState, controller} = props;
+  const [state, setState] = useState(initialState);
+
+  useEffect(
+    () => controller?.subscribe(() => setState({...controller.state})),
+    [controller]
+  );
+
+  return (
+    <ul>
+      {state.results.map((result) => (
+        <li key={result.uniqueId}>
+          <h3>{result.title}</h3>
+        </li>
+      ))}
+    </ul>
+  );
+};