+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 2.12.0 (2023-08-09)
+
+### Features
+
+- **atomic:** add skip link to grid results ([#3045](https://github.com/coveo/ui-kit/issues/3045)) ([049b420](https://github.com/coveo/ui-kit/commits/049b4201e72b91a49b2ec63ea2577fe3bed6a9e9))
+
+# 2.11.0 (2023-07-19)
+
+### Features
+
+- **atomic:** add the singular automatic facet component ([#3021](https://github.com/coveo/ui-kit/issues/3021)) ([ed45748](https://github.com/coveo/ui-kit/commits/ed4574841f8a130d6e78cfa80365f2129514f31c))
+
 # 2.10.0 (2023-06-29)
 
 ### Features